tests/wpt/tests/html/infrastructure/urls/base-url/document-base-url-changes-about-srcdoc-2.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>

<script>
// If document.open's behavior is modified to remove the url-rewriting behavior,
// then this test can be deleted (https://github.com/whatwg/html/issues/3989).
setup({ explicit_done: true });

window.start = childDocument => {
  const grandchildDocument =
    childDocument.getElementById('foo').contentDocument;

  test(t => {
    assert_equals(childDocument.URL, 'about:blank',
      'Child document starting URL');
    assert_equals(grandchildDocument.URL, 'about:srcdoc',
      'Grandchild document starting URL');
    const originalChildBaseURL = childDocument.baseURI;

    grandchildDocument.open("", "");
    // Verify that the document.open() trick worked: the grandchild should now
    // have the same url as the child, and have inherited the child's base url.
    assert_equals(grandchildDocument.URL, 'about:blank',
      'Grandchild document after document.open() trick');
    assert_equals(grandchildDocument.baseURI, originalChildBaseURL,
      'Grandchild base URL must match child base URL');

    // Give child a new base URL.
    const baseElement = childDocument.createElement('base');
    baseElement.href = get_host_info().REMOTE_ORIGIN;
    childDocument.head.append(baseElement);

    // Verify that changing the child's base url succeeded and did not affect
    // the grandchild's base url.
    const newChildBaseURL = childDocument.baseURI;
    assert_equals(grandchildDocument.URL, 'about:blank',
      'Grandchild document after child gets new base URL');
    assert_not_equals(newChildBaseURL, originalChildBaseURL,
      'Child base URL must change');
    assert_equals(grandchildDocument.baseURI, originalChildBaseURL,
      'Grandchild base URL must not change');
  });

  done();
};
</script>

<iframe src='about:blank'></iframe>

<script>
  window.onload = () => {
    let subframe_doc = document.querySelector('iframe').contentDocument;
    subframe_doc.body.innerHTML = '<iframe srcdoc="foo" id="foo"></iframe>';
    promise_test((test) => {
      return new Promise(async resolve => {
      // We need a timeout since the srcdoc frame takes time to setup and
      // doesn't fire a loadstop.
      test.step_timeout(resolve, 100);
      }).then(() => {
        assert_equals(
            subframe_doc.getElementById('foo').contentDocument.URL,
            'about:srcdoc');
        let script = subframe_doc.createElement('script');
        script.innerHTML = 'parent.start(document);';
        subframe_doc.body.appendChild(script);
      });
    }, "wrapper promise test for timeout.");
  };
</script>
